Got tests to run. Need to fix test errors and finish code cleanup.
