+++ junit5.bzl
@@ -1,0 +1,36 @@
+# Lovingly borrowed from
+# https://github.com/bmuschko/bazel-examples/blob/master/java/junit5-test/junit5.bzl.
+
+def java_junit5_test(name, srcs, test_package, deps = [], runtime_deps = [], **kwargs):
+    FILTER_KWARGS = [
+        "main_class",
+        "use_testrunner",
+        "args",
+    ]
+
+    for arg in FILTER_KWARGS:
+        if arg in kwargs.keys():
+            kwargs.pop(arg)
+
+    junit_console_args = []
+    if test_package:
+        junit_console_args += ["--select-package", test_package]
+    else:
+        fail("must specify 'test_package'")
+
+    native.java_test(
+        name = name,
+        srcs = srcs,
+        use_testrunner = False,
+        main_class = "org.junit.platform.console.ConsoleLauncher",
+        args = junit_console_args,
+        deps = deps + [
+            "@maven//:org_junit_jupiter_junit_jupiter_api",
+            "@maven//:org_junit_jupiter_junit_jupiter_params",
+            "@maven//:org_junit_jupiter_junit_jupiter_engine",
+        ],
+        runtime_deps = runtime_deps + [
+            "@maven//:org_junit_platform_junit_platform_console",
+        ],
+        **kwargs
+    )
