--- src/main/java/ca/ubc/cs/sanchom/AtmosView/AtmosView.java
@@ -1,17 +1,0 @@
-package ca.ubc.cs.sanchom.AtmosView;
-
-/**
- * The executable class for the Atmospheric Sounding Visualization.
- * @author Sancho McCann
- *
- */
-public class AtmosView {
-
-	public static void main(String args[]) {
-		MainFrame m = new MainFrame();
-		m.setVisible(true);
-		
-		return;
-	}
-	
-}
