+++ src/main/java/ca/ubc/cs/sanchom/atmosview/AtmosView.java
@@ -1,0 +1,16 @@
+package ca.ubc.cs.sanchom.atmosview;
+
+/**
+ * The executable class for the Atmospheric Sounding Visualization.
+ *
+ * @author Sancho McCann
+ */
+public class AtmosView {
+
+  public static void main(String args[]) {
+    MainFrame m = new MainFrame();
+    m.setVisible(true);
+
+    return;
+  }
+}
